member 모델은 courseId 추가한 것. 업데이트 페이지에 기존 회원 가입 정보 뜨도록 했고, 회원 정보가 변경되면 같이 변경될 수 있도록 설정.

+++ src/main/java/app/labs/profile/model/Profile.java
@@ -8,7 +8,7 @@ import lombok.ToString;
 @ToString //(exclude= {"profileImageData", "backgroundImageData"})
 
 public class Profile {
-	private Long profileId;
+	private int profileId;
     private String userId;
     private String userName;
     private String email;
